requirements.txt
fastapi==0.111.0
openai==1.26.0
pydantic==2.7.1
python-dotenv==1.0.1
Requests==2.31.0
